+++ NEWS.md
@@ -1,3 +1,8 @@
+# bandsfdp 1.1.0
+
+- Included function `gen_bound`, which computes an upper prediction bound on the FDP on any arbitrary set of hypotheses $R$.
+- Updated README to include an example of use.
+
 # bandsfdp 1.0.0
 
 -   Initial release of the bandsfdp package.